requirements.txt
discord.py==1.7.2
colorama==0.4.4
